// context/types.ts
export enum GetTodosActionTypes {
  GET_TODOS_REQUEST = 'GET_TODOS_REQUEST',
  GET_TODOS_SUCCESS = 'GET_TODOS_SUCCESS',
  GET_TODOS_FAIL = 'GET_TODOS_FAIL',
  GET_TODOS_RESET = 'GET_TODOS_RESET',
}

export enum GetTodoActionTypes {
  GET_TODO_REQUEST = 'GET_TODO_REQUEST',
  GET_TODO_SUCCESS = 'GET_TODO_SUCCESS',
  GET_TODO_FAIL = 'GET_TODO_FAIL',
  GET_TODO_RESET = 'GET_TODO_RESET',
}

export enum CreateTodoActionTypes {
  CREATE_TODO_REQUEST = 'CREATE_TODO_REQUEST',
  CREATE_TODO_SUCCESS = 'CREATE_TODO_SUCCESS',
  CREATE_TODO_FAIL = 'CREATE_TODO_FAIL',
  CREATE_TODO_RESET = 'CREATE_TODO_RESET',
}

export enum DeleteTodoActionTypes {
  DELETE_TODO_REQUEST = 'DELETE_TODO_REQUEST',
  DELETE_TODO_SUCCESS = 'DELETE_TODO_SUCCESS',
  DELETE_TODO_FAIL = 'DELETE_TODO_FAIL',
  DELETE_TODO_RESET = 'DELETE_TODO_RESET',
}

export enum UpdateTodoActionTypes {
  UPDATE_TODO_REQUEST = 'UPDATE_TODO_REQUEST',
  UPDATE_TODO_SUCCESS = 'UPDATE_TODO_SUCCESS',
  UPDATE_TODO_FAIL = 'UPDATE_TODO_FAIL',
  UPDATE_TODO_RESET = 'UPDATE_TODO_RESET',
}